Reorganize components folder
