+++ frontend/src/App.tsx
@@ -1,7 +1,7 @@
 import { ThemeProvider } from 'styled-components'
 import theme from './styles/theme'
 import GlobalStyles from './styles/globalStyles'
-import HealthEngagementPage from './components/HealthEngagementPage/HealthEngagementPage'
+import HealthEngagementPage from './components/pages/HealthEngagementPage/HealthEngagementPage'
 
 
 function App() {
