+++ frontend/src/components/layout/Footer.tsx
@@ -1,4 +1,4 @@
-import footerImage from '../assets/Footer.jpg';
+import footerImage from '../../assets/Footer.jpg';
 
 export default function Footer() {
   return <img className="fullWidthImage" src={footerImage} alt="Footer" />;
