+++ frontend/src/components/layout/Header.tsx
@@ -1,4 +1,4 @@
-import headerImage from '../assets/Header.jpg';
+import headerImage from '../../assets/Header.jpg';
 
 export default function Header() {
   return <img className="fullWidthImage" src={headerImage} alt="Header" />;
