+++ frontend/src/components/ui/ScoreDial/ScoreDial.styles.tsx
