+++ frontend/src/components/ui/ScoreDial/ScoreDial.tsx
